refactor: 각 도메인 디렉토리  구조 변경
