+++ src/main/java/com/theocean/fundering/domain/celebrity/domain/constant/CelebType.java
